refactor(broker-core): push ralph storage back to slack shell
