+++ broker-core/schema.ts
@@ -114,22 +114,6 @@ export interface TaskAssignmentAwaitingReplyInfo {
 
 // ─── Mappers ─────────────────────────────────────────────
 
-interface RalphCycleRow {
-  id: number;
-  started_at: string;
-  completed_at: string | null;
-  duration_ms: number | null;
-  ghost_agent_ids: string;
-  nudge_agent_ids: string;
-  idle_drain_agent_ids: string;
-  stuck_agent_ids: string;
-  anomalies: string;
-  anomaly_signature: string;
-  follow_up_delivered: number;
-  agent_count: number;
-  backlog_count: number;
-}
-
 function rowToAgent(row: AgentRow): AgentInfo {
   return {
     id: row.id,
@@ -374,27 +358,6 @@ function addObservabilityColumns(db: DatabaseSync): void {
     UPDATE agents
     SET idle_since = COALESCE(idle_since, last_seen)
     WHERE status = 'idle' AND idle_since IS NULL;
-  `);
-
-  db.exec(`
-    CREATE TABLE IF NOT EXISTS ralph_cycles (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      started_at TEXT NOT NULL,
-      completed_at TEXT,
-      duration_ms INTEGER,
-      ghost_agent_ids TEXT NOT NULL DEFAULT '[]',
-      nudge_agent_ids TEXT NOT NULL DEFAULT '[]',
-      idle_drain_agent_ids TEXT NOT NULL DEFAULT '[]',
-      stuck_agent_ids TEXT NOT NULL DEFAULT '[]',
-      anomalies TEXT NOT NULL DEFAULT '[]',
-      anomaly_signature TEXT NOT NULL DEFAULT '',
-      follow_up_delivered INTEGER NOT NULL DEFAULT 0,
-      agent_count INTEGER NOT NULL DEFAULT 0,
-      backlog_count INTEGER NOT NULL DEFAULT 0
-    );
-
-    CREATE INDEX IF NOT EXISTS idx_ralph_cycles_started
-      ON ralph_cycles(started_at);
   `);
 }
 
@@ -1735,85 +1698,6 @@ export class BrokerDB implements BrokerDBInterface {
     });
   }
 
-  // ─── Ralph cycles ────────────────────────────────────
-
-  recordRalphCycle(record: {
-    startedAt: string;
-    completedAt: string;
-    durationMs: number;
-    ghostAgentIds: string[];
-    nudgeAgentIds: string[];
-    idleDrainAgentIds: string[];
-    stuckAgentIds: string[];
-    anomalies: string[];
-    anomalySignature: string;
-    followUpDelivered: boolean;
-    agentCount: number;
-    backlogCount: number;
-  }): number {
-    const db = this.getDb();
-    const info = db
-      .prepare(
-        `INSERT INTO ralph_cycles (
-           started_at, completed_at, duration_ms,
-           ghost_agent_ids, nudge_agent_ids, idle_drain_agent_ids, stuck_agent_ids,
-           anomalies, anomaly_signature, follow_up_delivered,
-           agent_count, backlog_count
-         ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-      )
-      .run(
-        record.startedAt,
-        record.completedAt,
-        record.durationMs,
-        JSON.stringify(record.ghostAgentIds),
-        JSON.stringify(record.nudgeAgentIds),
-        JSON.stringify(record.idleDrainAgentIds),
-        JSON.stringify(record.stuckAgentIds),
-        JSON.stringify(record.anomalies),
-        record.anomalySignature,
-        record.followUpDelivered ? 1 : 0,
-        record.agentCount,
-        record.backlogCount,
-      );
-    return Number(info.lastInsertRowid);
-  }
-
-  getRecentRalphCycles(limit = 20): Array<{
-    id: number;
-    startedAt: string;
-    completedAt: string | null;
-    durationMs: number | null;
-    ghostAgentIds: string[];
-    nudgeAgentIds: string[];
-    idleDrainAgentIds: string[];
-    stuckAgentIds: string[];
-    anomalies: string[];
-    anomalySignature: string;
-    followUpDelivered: boolean;
-    agentCount: number;
-    backlogCount: number;
-  }> {
-    const db = this.getDb();
-    const rows = db
-      .prepare("SELECT * FROM ralph_cycles ORDER BY started_at DESC LIMIT ?")
-      .all(limit) as unknown as RalphCycleRow[];
-    return rows.map((row) => ({
-      id: row.id,
-      startedAt: row.started_at,
-      completedAt: row.completed_at,
-      durationMs: row.duration_ms,
-      ghostAgentIds: JSON.parse(row.ghost_agent_ids) as string[],
-      nudgeAgentIds: JSON.parse(row.nudge_agent_ids) as string[],
-      idleDrainAgentIds: JSON.parse(row.idle_drain_agent_ids) as string[],
-      stuckAgentIds: JSON.parse(row.stuck_agent_ids) as string[],
-      anomalies: JSON.parse(row.anomalies) as string[],
-      anomalySignature: row.anomaly_signature,
-      followUpDelivered: row.follow_up_delivered === 1,
-      agentCount: row.agent_count,
-      backlogCount: row.backlog_count,
-    }));
-  }
-
   repairThreadOwnership(): { releasedClaimCount: number; releasedAgentIds: string[] } {
     const db = this.getDb();
 
@@ -2243,7 +2127,7 @@ export class BrokerDB implements BrokerDBInterface {
     }
   }
 
-  private getDb(): DatabaseSync {
+  protected getDb(): DatabaseSync {
     if (!this.db) {
       throw new Error("BrokerDB not initialized — call initialize() first");
     }
